requirements.txt
googletrans2
selenium
pandas >= 0.25.3